third_party/WebKit/LayoutTests/editing/execCommand/insert-list-with-noneditable-content.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      'On^e two <span contenteditable="false" style="background-color: lightgray;">three</span> four<br>',
      'Fiv|e six.',
    '</div>'
  ].join(''),
  'insertUnorderedList',
  [
    '<div contenteditable>',
      '<ul>',
        '<li>',
          'On^e two <span contenteditable="false" style="background-color: lightgray;">three</span> four',
        '</li>',
        '<li>',
          'Fiv|e six.',
        '</li>',
      '</ul>',
    '</div>'
  ].join('')),
  'List creation across non-editable content should not crash.')
</script>
